+++ MANIFEST.in
@@ -1,3 +1,3 @@
 include requirements.txt
-include src/python_project_skeleton/*.ini
+include src/wunderbear/*.ini
 include *.sh
